refactor : aws패키지 코어로이동
