+++ core/src/main/java/kernel/maidlab/core/aws/dto/PresignedFileResponseDto.java
@@ -1,4 +1,4 @@
-package kernel.maidlab.domain.aws.dto;
+package kernel.maidlab.core.aws.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
