+++ core/src/main/java/kernel/maidlab/core/aws/service/RedisService.java
@@ -1,0 +1,5 @@
+package kernel.maidlab.core.aws.service;
+
+public interface RedisService {
+//	public void saveManagerList(String key, List<Manager> managers, long ttlMinutes);
+}
