+++ core/src/main/java/kernel/maidlab/core/aws/service/RedisServiceImpl.java
@@ -1,0 +1,5 @@
+package kernel.maidlab.core.aws.service;
+
+public class RedisServiceImpl {
+
+}
