--- domain/src/main/java/kernel/maidlab/domain/aws/service/RedisServiceImpl.java
@@ -1,5 +1,0 @@
-package kernel.maidlab.domain.aws.service;
-
-public class RedisServiceImpl {
-
-}
